Completed project reupload
is this a weather station project that I've built using Raspberry Pi 3 a few years ago, while I was learning more about the Linux environment and Phyton.  I didn't kept the original github I created for the project while I was working through the project, because I didn't thought I needed it anymore after a year or so since the completion of the project.   Thus I've reuploaded this project in a new repository.